@@ -1,19 +1,6 @@
-import '@babel/polyfill';
-// Import Component hooks before component definitions
-import './component-hooks';
-import Vue from 'vue';
-import './plugins/vuetify';
-import './plugins/vee-validate';
-import App from './App.vue';
-import router from './router';
-import store from '@/store';
-import './registerServiceWorker';
-import 'vuetify/dist/vuetify.min.css';
+import { createApp } from "vue";
+import App from "./App.vue";
+import router from "./router";
+import store from "./store";
 
-Vue.config.productionTip = false;
-
-new Vue({
-  router,
-  store,
-  render: (h) => h(App),
-}).$mount('#app');
+createApp(App).use(store).use(router).mount("#app");
