@@ -1,10 +1,3 @@
 module.exports = {
-  "presets": [
-    [
-      "@vue/cli-plugin-babel/preset",
-      {
-        "useBuiltIns": "entry"
-      }
-    ]
-  ]
-}
+  presets: ["@vue/cli-plugin-babel/preset"],
+};
